[#53] refactor: product 모듈에서 campaign 의존성 제거
