+++ live-commerce-core/src/main/java/org/giglab/live/commerce/core/campaign/infrastructure/persistence/CampaignQueryRepository.java
@@ -14,8 +14,8 @@ import org.giglab.live.commerce.core.campaign.application.dto.CampaignListQuery;
 import org.giglab.live.commerce.core.campaign.application.dto.CampaignProductDto;
 import org.giglab.live.commerce.core.campaign.application.dto.CampaignSummary;
 import org.giglab.live.commerce.core.campaign.application.dto.GetCampaignResult;
-import org.giglab.live.commerce.core.campaign.application.dto.ProductLinkedCampaignDto;
 import org.giglab.live.commerce.core.global.jpa.entity.types.YnType;
+import org.giglab.live.commerce.core.shared.ProductLinkedCampaignDto;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
