+++ live-commerce-core/src/main/java/org/giglab/live/commerce/core/product/application/dto/GetProductLinkedCampaignsResult.java
@@ -1,6 +1,6 @@
 package org.giglab.live.commerce.core.product.application.dto;
 
 import java.util.List;
-import org.giglab.live.commerce.core.campaign.application.dto.ProductLinkedCampaignDto;
+import org.giglab.live.commerce.core.shared.ProductLinkedCampaignDto;
 
 public record GetProductLinkedCampaignsResult(List<ProductLinkedCampaignDto> campaigns) {}
