+++ live-commerce-core/src/main/java/org/giglab/live/commerce/core/product/application/usecase/GetProductLinkedCampaignsUseCase.java
@@ -1,8 +1,8 @@
 package org.giglab.live.commerce.core.product.application.usecase;
 
 import lombok.RequiredArgsConstructor;
-import org.giglab.live.commerce.core.campaign.application.port.persistence.CampaignQueryPort;
 import org.giglab.live.commerce.core.product.application.dto.GetProductLinkedCampaignsResult;
+import org.giglab.live.commerce.core.product.application.port.bridge.ProductCampaignAppPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,9 +11,10 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class GetProductLinkedCampaignsUseCase {
 
-  private final CampaignQueryPort campaignQueryPort;
+  private final ProductCampaignAppPort productCampaignAppPort;
 
   public GetProductLinkedCampaignsResult execute(Long productId) {
-    return new GetProductLinkedCampaignsResult(campaignQueryPort.findByProductId(productId));
+    return new GetProductLinkedCampaignsResult(
+        productCampaignAppPort.getLinkedCampaignList(productId));
   }
 }
