+++ live-commerce-core/src/main/java/org/giglab/live/commerce/core/shared/ProductLinkedCampaignDto.java
@@ -1,4 +1,4 @@
-package org.giglab.live.commerce.core.campaign.application.dto;
+package org.giglab.live.commerce.core.shared;
 
 import java.time.LocalDateTime;
 import org.giglab.live.commerce.core.campaign.domain.entity.types.BroadcastStatusType;
